Move king castle classes to separate module
